# docs/swagger.yaml
basePath: /
definitions:
  github_com_williamkoller_go-ddd-api_internal_user_domain.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:3003
info:
  contact: {}
  description: API com DDD, Auth e CRUD de usuário
  title: Go DDD API
  version: 0.0.1
paths:
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_williamkoller_go-ddd-api_internal_user_domain.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar todos os usuários
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deletar usuário
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_williamkoller_go-ddd-api_internal_user_domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar usuário
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais de login
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login do usuário
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Usuário a ser registrado
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_williamkoller_go-ddd-api_internal_user_domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_williamkoller_go-ddd-api_internal_user_domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar usuário
      tags:
      - Users
schemes:
- http,
- https
swagger: "2.0"